+++ src/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 v.4.1.2
 * [bug] Very unusual bug when copying taxonomy template files (php-fpm with umask 002, could copy but not reading due to permissions, added an extra check in Copy_Template_Taxonomy::copy_taxonomy_template()
+* [technical] WP functions standardization and escaping (thanks to the new WP plugin check https://wordpress.org/plugins/plugin-check/)
 
 v.4.1.1
 * [feature] Notifications messages translated.